Added client-console.sh and client-start.sh scripts.
These scripts will help the startup of the frontend development environment.
You can execute client-console.sh and it will launch the frontend development
server inside a screen session and it will attach to it.

If you just want to launch the frontend development server, then you can execute client-start.sh which will launch the server in detached screen session. If you then want to see the messages which appears in the screen, you can attach to the session using client-console.sh script.