Fixed SSL cert generation for iOS 13 and macOS 10.15. (#46)
Apple has changed requirements for trusted certificates in iOS 13 and macOS 10.15 (see HT210176). The certificates must be valid for maximum 825 and they must have subjectAltName and extendedKeyUsage set for serverAuth. These changes will keep the self-signed certificates in a shape that those services can be contacted using iOS and macOS now and in the near future.