Merge master to release v0.10.2-rc3 (#57)
* This will update the tests to match the new title. (#27)

This will also enable the testing to be done within qvain-test.csc.fi.

* Added missing LICENSE (#28)

The MIT license has been added, the copyright is Ministry of Education and Culture, Finland.

* CSCQVAIN-348: Read stats API key from secrets (#29)

* Fixed local builds. (#30)

* Enable builds on macOS for Linux platform. (#31)

Enable builds on macOS for Linux platform.
Added helpers for stable and test deployments.

* Fixed builds. (#32)

The provisioning fails when upgrading environment without those flags.

* CSCQVAIN-373: Add lookup API key to qvain-api environment (#33)

* Fixed builds for qvain-test deployment (#34)

* The scripts were missing set -e. (#35)

* CSCQVAIN-485: Fixed node version to 10.x series. (#37)

* CSCQVAIN-485: set default nodejs to 10.x. (#38)

We need to call nvm use 10 to ensure that the nodejs 10.x is selected, as it can be that there is a newer installed and that is default.

* CSCQVAIN-485: fixed typo. (#39)

* CSCQVAIN-485: Fix another typo. (#40)

We need to rewrite the whole ansible roles at some point, this is just horrible and nasty looking stuff..

* CSCQVAIN-485: the default alias is broken with nvm. (#41)

It must be called in order to have it up and running.

* CSCFAIRMETA-266: Fixed nginx configuration. (#42)

* CSCFAIRMETA-266: Fixed SSL headers for Matomo. (#43)

* CSCFAIRMETA-324: Fixed broken Vagrantfile. (#44)

There was broken command for Vagrant inline script part.
Also it was missing failure handling.

* CSCFAIRMETA-291: Increase nginx proxy timeout (#45)

* Fixed SSL cert generation for iOS 13 and macOS 10.15. (#46)

Apple has changed requirements for trusted certificates in iOS 13 and macOS 10.15 (see HT210176). The certificates must be valid for maximum 825 and they must have subjectAltName and extendedKeyUsage set for serverAuth. These changes will keep the self-signed certificates in a shape that those services can be contacted using iOS and macOS now and in the near future.

* CSCFAIRMETA-334: Create reinstall_admin_deps.sh script (#48)

* CSCFAIRMETA-334: Create reinstall_admin_deps.sh script

- Needed for alignment purposes with metax-ops & etsin-ops
- Simple script defining the shell as bash, set -e for error handling, and then relevant deployment commands

* CSCFAIRMETA-334: Make the .sh script executable

- Previously, the reinstall_admin_deps.sh script was not executable, this commit makes it executable
- Script made executable using a chmod command

* Enabled Travis CI for this repository. (#49)

* Merge next to master v0.10.2-rc1 (#50)

* This will update the tests to match the new title. (#27)

This will also enable the testing to be done within qvain-test.csc.fi.

* Added missing LICENSE (#28)

The MIT license has been added, the copyright is Ministry of Education and Culture, Finland.

* CSCQVAIN-348: Read stats API key from secrets (#29)

* Fixed local builds. (#30)

* Enable builds on macOS for Linux platform. (#31)

Enable builds on macOS for Linux platform.
Added helpers for stable and test deployments.

* Fixed builds. (#32)

The provisioning fails when upgrading environment without those flags.

* CSCQVAIN-373: Add lookup API key to qvain-api environment (#33)

* Fixed builds for qvain-test deployment (#34)

* The scripts were missing set -e. (#35)

* CSCQVAIN-485: Fixed node version to 10.x series. (#37)

* CSCQVAIN-485: set default nodejs to 10.x. (#38)

We need to call nvm use 10 to ensure that the nodejs 10.x is selected, as it can be that there is a newer installed and that is default.

* CSCQVAIN-485: fixed typo. (#39)

* CSCQVAIN-485: Fix another typo. (#40)

We need to rewrite the whole ansible roles at some point, this is just horrible and nasty looking stuff..

* CSCQVAIN-485: the default alias is broken with nvm. (#41)

It must be called in order to have it up and running.

* CSCFAIRMETA-266: Fixed nginx configuration. (#42)

* CSCFAIRMETA-266: Fixed SSL headers for Matomo. (#43)

* CSCFAIRMETA-324: Fixed broken Vagrantfile. (#44)

There was broken command for Vagrant inline script part.
Also it was missing failure handling.

* CSCFAIRMETA-291: Increase nginx proxy timeout (#45)

* Fixed SSL cert generation for iOS 13 and macOS 10.15. (#46)

Apple has changed requirements for trusted certificates in iOS 13 and macOS 10.15 (see HT210176). The certificates must be valid for maximum 825 and they must have subjectAltName and extendedKeyUsage set for serverAuth. These changes will keep the self-signed certificates in a shape that those services can be contacted using iOS and macOS now and in the near future.

* CSCFAIRMETA-334: Create reinstall_admin_deps.sh script (#48)

* CSCFAIRMETA-334: Create reinstall_admin_deps.sh script

- Needed for alignment purposes with metax-ops & etsin-ops
- Simple script defining the shell as bash, set -e for error handling, and then relevant deployment commands

* CSCFAIRMETA-334: Make the .sh script executable

- Previously, the reinstall_admin_deps.sh script was not executable, this commit makes it executable
- Script made executable using a chmod command

* Enabled Travis CI for this repository. (#49)

Co-authored-by: Juhapekka Piiroinen <jp@1337.fi>
Co-authored-by: Jori Niemi <3295718+tahme@users.noreply.github.com>

* Fixed unicode space issues. (#51)

* Remove extra "repo_branch: next" from stable inventory (#52)

* Use git 2.9 for building on remote machine (#54)

* Fixed makefile. (#55)

Co-authored-by: Juhapekka Piiroinen <jp@1337.fi>
Co-authored-by: Jori Niemi <3295718+tahme@users.noreply.github.com>